Fix miscalculated mmap args for massive dynamic-space-size
With mark-region's 128b card size it is quite possible to need a card table
of 2 gigacards. This happens with a dynamic-space-size slightly over 128GiB.
(Exactly 128GiB = 1 gigacard). So this how the crash went: 2 gigacards uses
a 31-bit mask, which is #x7fffffff and 1+ that looks like the most negative
32-bit int. This caused 'extra_above' in coreparse_alloc_space to add -huge
to 'size' which made the call to os_allloc_gc_space receive a bad size
such as 0xffffffffzzzzzzzz which produces ENOMEM.

I don't know how to write a regression test for a "typical" setup though my
particular setup allowed a heap size of 130GiB other than for the bad math.

+++ src/runtime/arm64-arch.c
@@ -361,7 +361,7 @@ void gcbarrier_patch_code(void* where, int nbits)
 os_vm_address_t coreparse_alloc_space(int space_id, int attr,
                                       os_vm_address_t addr, os_vm_size_t size)
 {
-    __attribute__((unused)) int extra_request = 0;
+    __attribute__((unused)) long extra_request = 0;
 #ifdef LISP_FEATURE_IMMOBILE_SPACE
     if (space_id == IMMOBILE_TEXT_CORE_SPACE_ID) {
         extra_request = ALIEN_LINKAGE_SPACE_SIZE;
